Разделяет редьюсер
